refactor[#226] : 단순한 컴포넌트는 페이지 레이어, setting.ts 에서 사용 할 컴포넌트들 정의
