+++ src/entities/setting/ui/index.ts
@@ -1,2 +1,1 @@
-export * from "./VersionInfo";
 export * from "./Account";
