+++ src/features/setting/ui/index.ts
@@ -1,8 +1,2 @@
-export * from "./EditMyInfo";
-export * from "./AccountManagement";
-export * from "./MyActivity";
-export * from "./Notification";
-export * from "./TermOfService";
-export * from "./Report";
-export * from "./LogoutButton";
 export * from "./Account";
+export * from "./setting";
