+++ src/features/setting/ui/setting.tsx
@@ -1,8 +1,13 @@
+/**
+ * 해당 파일은 /my-page/setting 에서 사용 되는 컴포넌트를 모아둔 파일입니다.
+ * page 레이어에서 정의 되기엔 비즈니스 로직이나 도메인 로직이 존재하는 컴포넌트 들을 모아뒀습니다.
+ */
 import { useSnackBar } from "@/shared/lib";
+import { useModal } from "@/shared/lib";
 import { Snackbar } from "@/shared/ui/snackbar";
+import { LogoutModal } from "./_LogoutModal";
 import { settingClassName } from "./setting.styles";
 
-// 문의/제안
 export const Report = () => {
   const officialEmail = "mungwithme@gmail.com";
 
@@ -26,6 +31,18 @@ export const Report = () => {
     <button onClick={handleClick} className={settingClassName}>
       <p>문의/제안</p>
       <span className="text-grey-700 body-2">mungwithme@gmail.com</span>
+    </button>
+  );
+};
+
+export const LogoutButton = () => {
+  const { handleOpen: handleOpenLogoutModal, onClose } = useModal(() => (
+    <LogoutModal onCloseLogoutModal={onClose} />
+  ));
+
+  return (
+    <button className={settingClassName} onClick={handleOpenLogoutModal}>
+      로그아웃
     </button>
   );
 };
